refactor: HomeUI 컴포넌트화
